Fix Yarn, Move to TypeScript for Frontend
