+++ package.json
@@ -4,11 +4,18 @@
 	"private": true,
 	"dependencies": {
 		"@apollo/client": "^3.3.6",
+		"@apollo/link-context": "^2.0.0-beta.3",
 		"@auth0/auth0-react": "^1.2.0",
 		"@reduxjs/toolkit": "^1.5.0",
 		"@testing-library/jest-dom": "^5.11.4",
 		"@testing-library/react": "^11.1.0",
 		"@testing-library/user-event": "^12.1.10",
+		"@types/jest": "^26.0.19",
+		"@types/node": "^14.14.19",
+		"@types/react": "^17.0.0",
+		"@types/react-dom": "^17.0.0",
+		"@types/react-router-dom": "^5.1.7",
+		"@types/styled-components": "^5.1.7",
 		"apollo-link-context": "^1.0.20",
 		"apollo-link-ws": "^1.0.20",
 		"dayjs": "^1.9.8",
@@ -22,6 +29,7 @@
 		"react-scripts": "4.0.1",
 		"styled-components": "^5.2.1",
 		"subscriptions-transport-ws": "^0.9.18",
+		"typescript": "^4.1.3",
 		"web-vitals": "^0.2.4"
 	},
 	"scripts": {
@@ -49,5 +57,9 @@
 			"last 1 firefox version",
 			"last 1 safari version"
 		]
+	},
+	"devDependencies": {
+		"@types/react-loader-spinner": "^3.1.3",
+		"@types/react-redux": "^7.1.15"
 	}
 }
