+++ src/components/JobDetailsForm.tsx
@@ -1,11 +1,10 @@
 import React from "react";
 import styled from "styled-components";
-import { Formik, Form, Field, FieldArray, ErrorMessage } from "formik";
+import { Formik, Form, Field, ErrorMessage } from "formik";
 
 import Section from "./styledComponents/Section";
 import SectionHeader from "./styledComponents/SectionHeader";
 import SectionElement from "./styledComponents/SectionElement";
-
 const SubmitButton = styled.button`
 	width: 100%;
 	background-color: rgba(0, 200, 0, 0.7);
@@ -21,6 +20,71 @@ const SubmitButton = styled.button`
 		cursor: pointer;
 	}
 `;
+// TODO: vvvvvv
+const deleteJob = (id: any) => {
+	console.log(`deleting job ${id}!`);
+};
+
+const CustomerInfoSection = styled(Section)`
+	display: grid;
+	grid-template-columns: 1fr 1fr 1fr;
+	grid-template-areas:
+		"header header header"
+		"content content content";
+	@media (max-width: 500px) {
+		grid-template-columns: 1fr 1fr;
+		grid-template-areas:
+			"header header "
+			"content content";
+	}
+`;
+
+const JobDetailsSection = styled.div`
+	display: grid;
+	grid-template-columns: 1fr 1fr 1fr;
+	gridarea: "machineDetails";
+	@media (max-width: 500px) {
+		grid-template-columns: 1fr 1fr;
+	}
+`;
+
+export const CustomerInfo = () => {
+	return (
+		<CustomerInfoSection>
+			<SectionHeader>Customer Information</SectionHeader>
+			<SectionElement>
+				<label htmlFor="firstName">First Name</label>
+				<Field type="firstName" name="firstName" />
+				<ErrorMessage name="firstName" component="div" />
+			</SectionElement>
+			<SectionElement>
+				<label htmlFor="lastName">Last Name</label>
+				<Field type="lastName" name="lastName" />
+				<ErrorMessage name="lastName" component="div" />
+			</SectionElement>
+			<SectionElement>
+				<label htmlFor="email">E-Mail</label>
+				<Field type="email" name="email" />
+				<ErrorMessage name="email" component="div" />
+			</SectionElement>
+			<SectionElement>
+				<label htmlFor="city">City</label>
+				<Field type="city" name="city" />
+				<ErrorMessage name="city" component="div" />
+			</SectionElement>
+			<SectionElement>
+				<label htmlFor="district">District</label>
+				<Field type="district" name="district" />
+				<ErrorMessage name="district" component="div" />
+			</SectionElement>
+			<SectionElement>
+				<label htmlFor="postCode">Postcode</label>
+				<Field type="postcode" name="postcode" />
+				<ErrorMessage name="postcode" component="div" />
+			</SectionElement>
+		</CustomerInfoSection>
+	);
+};
 
 const YearSelection = () => {
 	let years = [];
@@ -28,78 +92,19 @@ const YearSelection = () => {
 		years.push(year);
 	}
 
-	return years.map((year) => (
-		<option key={year} value={year}>
-			{year}
-		</option>
-	));
-};
-
-export const CustomerInfo = () => {
 	return (
-		<Section
-			style={{
-				display: "grid",
-				gridTemplateColumns: "1fr 1fr 1fr",
-				alignItems: "flex-start",
-				gridTemplateAreas: `
-                        'header header header'
-                        'content content content'
-                    `,
-			}}
-		>
-			<SectionHeader>Customer Information</SectionHeader>
-			<SectionElement>
-				<label htmlFor="firstName">First Name</label>
-				<Field type="firstName" name="firstName" />
-				<ErrorMessage
-					style={{ color: "red" }}
-					name="firstName"
-					component="div"
-				/>
-			</SectionElement>
-			<SectionElement>
-				<label htmlFor="lastName">Last Name</label>
-				<Field type="lastName" name="lastName" />
-				<ErrorMessage
-					style={{ color: "red" }}
-					name="lastName"
-					component="div"
-				/>
-			</SectionElement>
-			<SectionElement>
-				<label htmlFor="email">E-Mail</label>
-				<Field type="email" name="email" />
-				<ErrorMessage style={{ color: "red" }} name="email" component="div" />
-			</SectionElement>
-			<SectionElement>
-				<label htmlFor="city">City</label>
-				<Field type="city" name="city" />
-				<ErrorMessage style={{ color: "red" }} name="city" component="div" />
-			</SectionElement>
-			<SectionElement>
-				<label htmlFor="district">District</label>
-				<Field type="district" name="district" />
-				<ErrorMessage
-					style={{ color: "red" }}
-					name="district"
-					component="div"
-				/>
-			</SectionElement>
-			<SectionElement>
-				<label htmlFor="postCode">Postcode</label>
-				<Field type="postcode" name="postcode" />
-				<ErrorMessage
-					style={{ color: "red" }}
-					name="postcode"
-					component="div"
-				/>
-			</SectionElement>
-		</Section>
+		<Field style={{ height: "2.5rem" }} as="select" name="year">
+			{years.map((year) => (
+				<option key={year} value={year}>
+					{year}
+				</option>
+			))}
+		</Field>
 	);
 };
 
-export const JobDetails = ({ values, handleChange }) => (
+// TODO: vvvvvvvvvvvvv
+export const JobDetails = ({ values, handleChange }: any) => (
 	<Section
 		style={{
 			display: "grid",
@@ -107,44 +112,35 @@ export const JobDetails = ({ values, handleChange }) => (
 			gridTemplateAreas: `
 			'header '
 			'machineDetails'
-                        'contentt '
+          'contentt '
                     `,
 		}}
 	>
-		<div
-			style={{
-				display: "grid",
-				gridTemplateColumns: "1fr 1fr 1fr",
-				gridArea: "machineDetails",
-				alignItems: "flex-start",
-			}}
-		>
+		<JobDetailsSection>
 			<SectionElement>
 				<label htmlFor="make">Make</label>
 				<Field type="make" name="make" />
-				<ErrorMessage style={{ color: "red" }} name="make" component="div" />
+				<ErrorMessage name="make" component="div" />
 			</SectionElement>
 
 			<SectionElement>
 				<label htmlFor="model">Model</label>
 				<Field type="model" name="model" />
-				<ErrorMessage style={{ color: "red" }} name="model" component="div" />
+				<ErrorMessage name="model" component="div" />
 			</SectionElement>
 
 			<SectionElement>
 				<label htmlFor="year">Year</label>
 
-				<Field style={{ height: "2.5rem" }} as="select" name="year">
-					<YearSelection />
-				</Field>
+				<YearSelection />
 
-				<ErrorMessage style={{ color: "red" }} name="year" component="div" />
+				<ErrorMessage name="year" component="div" />
 			</SectionElement>
 
 			<SectionElement>
 				<label htmlFor="serial">Serial</label>
 				<Field type="serial" name="serial" />
-				<ErrorMessage style={{ color: "red" }} name="serial" component="div" />
+				<ErrorMessage name="serial" component="div" />
 			</SectionElement>
 
 			<SectionElement>
@@ -155,26 +151,20 @@ export const JobDetails = ({ values, handleChange }) => (
 					type="status"
 					name="status"
 				>
-					<option default value="not-started">
-						Not Started
-					</option>
+					<option value="not-started">Not Started</option>
 					<option value="await">Awaiting Parts</option>
 					<option value="fuck">FUCK!</option>
 					<option value="completed">Completed</option>
 				</Field>
-				<ErrorMessage style={{ color: "red" }} name="status" component="div" />
+				<ErrorMessage name="status" component="div" />
 			</SectionElement>
 
 			<SectionElement>
 				<label htmlFor="labourHours">Labour (hrs)</label>
 				<Field type="labourHours" name="labourHours" />
-				<ErrorMessage
-					style={{ color: "red" }}
-					name="labourHours"
-					component="div"
-				/>
+				<ErrorMessage name="labourHours" component="div" />
 			</SectionElement>
-		</div>
+		</JobDetailsSection>
 		<SectionHeader>Job Details</SectionHeader>
 
 		<div
@@ -189,13 +179,13 @@ export const JobDetails = ({ values, handleChange }) => (
 				onChange={handleChange}
 				name="todo"
 				value={values.todo}
-				rows="7"
+				rows={7}
 			></textarea>
 			<ErrorMessage name="todo" component="div" />
 
 			<label htmlFor="done">Done</label>
 			<textarea
-				rows="7"
+				rows={7}
 				onChange={handleChange}
 				name="done"
 				value={values.done}
@@ -204,8 +194,8 @@ export const JobDetails = ({ values, handleChange }) => (
 		</div>
 	</Section>
 );
-
-export const InternalUse = () => (
+// TODO: vvvvvvvvv
+export const InternalUse = ({ id }: any) => (
 	<Section
 		style={{
 			display: "grid",
@@ -218,6 +208,14 @@ export const InternalUse = () => (
 	>
 		<SectionHeader>Internal Use</SectionHeader>
 		stuff
+		<button
+			type="button"
+			onClick={() => {
+				deleteJob(id);
+			}}
+		>
+			Delete Job
+		</button>
 	</Section>
 );
 
@@ -235,60 +233,43 @@ export const Parts = () => {
 		>
 			<SectionHeader>Parts</SectionHeader>
 			stuff
-			<table
-				style={{
-					width: "100%",
-					textAlign: "center",
-				}}
-			>
-				<thead>
-					<tr style={{ fontSize: "1.1rem" }}>
-						<td>Status</td>
-						<td>ID</td>
-						<td>Client</td>
-						<td>Created</td>
-					</tr>
-				</thead>
-				<tbody>
-					<tr>
-						<td colSpan={4}>No Parts</td>
-					</tr>
-				</tbody>
-			</table>
 		</Section>
 	);
 };
 
-export const Submit = ({ isSubmitting }) => (
+// TODO: vvvvvvvv
+export const Submit = ({ isSubmitting }: any) => (
 	<div style={{ width: "100%", textAlign: "center" }}>
 		<SubmitButton type="submit" disabled={isSubmitting}>
-			Add Job
+			Save Job
 		</SubmitButton>
 	</div>
 );
 
-export const NewJobForm = () => {
+//TODO:
+export const JobDetailsForm = ({ data }: any) => {
 	return (
 		<Formik
 			initialValues={{
-				firstName: "",
-				lastName: "",
-				email: "",
-				city: "",
-				district: "",
-				postcode: "",
-				date: "",
-				todo: "",
-				done: "",
-				make: "",
-				model: "",
-				year: "",
-				serial: "",
-				status: "",
-				labourHours: "",
+				firstName: `${data?.firstname}`,
+				lastName: `${data?.lastname}`,
+				email: `${data?.email}`,
+				city: `${data?.city}`,
+				district: `${data?.district}`,
+				postcode: `${data?.postcode}`,
+				date: `${data?.date}`,
+				todo: `${data?.todo}`,
+				done: `${data?.done}`,
+				status: `${data?.status}`,
+				assigned: `${data?.assigned}`,
+				make: `${data?.make}`,
+				model: `${data?.model}`,
+				year: `${data?.year}`,
+				serial: `${data?.serial}`,
+				labourHours: `${data?.labourHours}`,
 			}}
 			validate={(values) => {
-				const errors = {};
+				const errors: any = {};
 				if (!values.email) {
 					errors.email = "Required";
 				} else if (
@@ -316,7 +297,7 @@ export const NewJobForm = () => {
 					<JobDetails values={values} handleChange={handleChange} />
 
 					<Parts />
-					<InternalUse />
+					<InternalUse id={data?._id} />
 					<Submit isSubmitting={isSubmitting} />
 				</Form>
 			)}
@@ -324,4 +305,4 @@ export const NewJobForm = () => {
 	);
 };
 
-export default NewJobForm;
+export default JobDetailsForm;
