+++ src/components/styledComponents/SectionElement.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-
-const SectionElement = ({ children }) => {
+//TODO: vvvvvvvvvvvvvvv
+const SectionElement = ({ children }: any) => {
 	return <div style={{ display: "grid", margin: "0.25rem" }}>{children}</div>;
 };
 export default SectionElement;
