+++ src/features/authData.tsx
@@ -2,18 +2,18 @@ import { createSlice } from "@reduxjs/toolkit";
 const initialState = { isLoggedIn: false };
 
 const authSlice = createSlice({
-  name: "authSlice",
-  initialState: initialState,
-  reducers: {
-    doLogin(state) {
-      console.log('loggin in')
-      state.isLoggedIn = true;
-    },
+	name: "authSlice",
+	initialState: initialState,
+	reducers: {
+		doLogin(state) {
+			console.log("loggin in");
+			state.isLoggedIn = true;
+		},
 
-    doLogout(state) {
-      state.isLoggedIn = false;
-    },
-  },
+		doLogout(state) {
+			state.isLoggedIn = false;
+		},
+	},
 });
 
 export const { doLogin, doLogout } = authSlice.actions;
